refactor(-): :hammer: move validators dieta especial
