+++ sme_sigpae_api/dieta_especial/solicitacao_dieta_especial/api/validators.py
@@ -36,15 +36,6 @@ def somente_digitos(codigo_eol):
         raise serializers.ValidationError("Deve ter somente dígitos")
 
 
-def edital_ja_existe_protocolo(editais, quantidade_editais_enviados=1):
-    if len(editais) > 0 and quantidade_editais_enviados > 1:
-        str_editais = ", ".join(str(edital["numero"]) for edital in editais)
-        raise serializers.ValidationError(
-            f"Já existe um protocolo padrão com esse nome para os editais: {str_editais}."
-        )
-    raise serializers.ValidationError("Já existe um protocolo padrão com esse nome.")
-
-
 class AlunoSerializerValidator(serializers.Serializer):
     codigo_eol = serializers.CharField(max_length=7, validators=[somente_digitos])
     nome = serializers.CharField(max_length=100)
